parse from float to int

+++ main.py
@@ -9,7 +9,7 @@ def newton_symbol(lon):
     from math import factorial
     n = int(lon[0])
     k = int(lon[1])
-    return factorial(n)/(factorial(k) * factorial(n-k))
+    return int(factorial(n)/(factorial(k) * factorial(n-k)))
 
 
 print(newton_symbol(list_of_numbers))
